backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi>=2026.2.25",
    "fastapi>=0.135.1",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "supabase>=2.28.3",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "uvicorn[standard]>=0.42.0",
]
